requirements.txt
fastapi
uvicorn
pydantic
speechrecognition
pydub
python-dotenv
llama-index-core
llama-index-llms-groq
llama-index-vector-stores-pinecone
llama-index-embeddings-huggingface
pinecone
pandas
python-multipart